拆完元件~修bug~;寫有關後端的功能
