+++ next-client/pages/user/product-like.js
@@ -8,8 +8,8 @@ import {
   faCaretDown,
   faCaretRight,
 } from '@fortawesome/free-solid-svg-icons'
-import UserSideBar from './user-side-bar'
-import UserSideBarMobile from '../../components/user/user-side-bar-mobile'
+import UserSideBar from '@/components/user/user-side-bar'
+import UserSideBarMobile from '@/components/user/user-side-bar-mobile'
 export default function Detail() {
   // const { RangePicker } = DatePicker
   return (
