+++ next-client/styles/user/register.module.scss
@@ -4,3 +4,8 @@
     }
 
   }
+
+  input.hide-arrow::-webkit-outer-spin-button,
+  input.hide-arrow::-webkit-inner-spin-button {
+    -webkit-appearance: none;
+  }
